# api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookmarkHive API",
        "description": "Decentralized social bookmarking service based on the Activity Pub protocol.",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://bookmarkhive.test",
            "description": "API Dev Server"
        },
        {
            "url": "https://api2.bookmarkhive.test",
            "description": "API 2 Dev Server"
        }
    ],
    "paths": {
...
        "/users/me/bookmarks": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "List own bookmarks",
                "description": "Returns a paginated collection of bookmarks owned by the authenticated user. Supports filtering by tags and cursor-based pagination. Default page size is 24 items, ordered by creation date (newest first).",
                "operationId": "listOwnBookmarks",
                "parameters": [
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Comma-separated list of tag slugs to filter by",
                        "schema": {
                            "type": "string",
                            "example": "tag-one,tag-two"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query string",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "after",
                        "in": "query",
                        "description": "Cursor for pagination - bookmark ID to fetch results after",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookmarks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collection": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BookmarkShowPrivate"
                                            }
                                        },
                                        "prevPage": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "nextPage": {
                                            "description": "URL for next page if available",
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "bookmark_list": {
                                        "summary": "Paginated list of bookmarks",
                                        "value": {
                                            "collection": [
                                                {
                                                    "id": "01234567-89ab-cdef-0123-456789abcdef",
                                                    "createdAt": "2024-01-01T12:00:00+00:00",
                                                    "title": "Example Bookmark",
                                                    "url": "https://example.com",
                                                    "domain": "example.com",
                                                    "account": {
                                                        "@iri": "https://bookmarkhive.test/profile/johndoe",
                                                        "username": "johndoe",
                                                        "instance": "bookmarkhive.test"
                                                    },
                                                    "isPublic": true,
                                                    "tags": [
                                                        {
                                                            "name": "Bookmarking",
                                                            "slug": "bookmarking",
                                                            "@iri": "https://bookmarkhive.test/users/me/tags/bookmarking"
                                                        }
                                                    ],
                                                    "instance": "bookmarkhive.test",
                                                    "@iri": "https://bookmarkhive.test/users/me/bookmarks/01234567-89ab-cdef-0123-456789abcdef"
                                                }
                                            ],
                                            "prevPage": null,
                                            "nextPage": "https://bookmarkhive.test/users/me/bookmarks?after=01234567-89ab-cdef-0123-456789abcdef",
                                            "total": null
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Create a new bookmark",
                "description": "Creates a new bookmark. If a bookmark with the same URL already exists, the previous one will be marked as outdated.",
                "operationId": "createBookmark",
                "requestBody": {
                    "description": "Bookmark data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "url"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Bookmark title",
                                        "type": "string"
                                    },
                                    "url": {
                                        "description": "Bookmark URL",
                                        "type": "string",
                                        "format": "uri"
                                    },
                                    "isPublic": {
                                        "description": "Whether the bookmark is public",
                                        "type": "boolean",
                                        "default": false
                                    },
                                    "tags": {
                                        "description": "Array of tag IRIs (must be valid IRIs pointing to existing tags owned by the user)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "iri",
                                            "example": "https://bookmarkhive.test/users/me/tags/web-development"
                                        }
                                    },
                                    "mainImage": {
                                        "description": "IRI of main image FileObject",
                                        "type": "string",
                                        "format": "iri",
                                        "nullable": true
                                    },
                                    "archive": {
                                        "description": "IRI of archive FileObject",
                                        "type": "string",
                                        "format": "iri",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                },
                                "examples": {
                                    "created_bookmark": {
                                        "summary": "Successfully created bookmark",
                                        "value": {
                                            "id": "01234567-89ab-cdef-0123-456789abcdef",
                                            "createdAt": "2024-01-01T12:00:00+00:00",
                                            "title": "Example Bookmark",
                                            "url": "https://example.com",
                                            "domain": "example.com",
                                            "account": {
                                                "@iri": "https://bookmarkhive.test/profile/johndoe",
                                                "username": "johndoe",
                                                "instance": "bookmarkhive.test"
                                            },
                                            "isPublic": false,
                                            "tags": [
                                                {
                                                    "name": "Bookmarking",
                                                    "slug": "bookmarking",
                                                    "@iri": "https://bookmarkhive.test/users/me/tags/bookmarking"
                                                }
                                            ],
                                            "instance": "bookmarkhive.test",
                                            "@iri": "https://bookmarkhive.test/users/me/bookmarks/01234567-89ab-cdef-0123-456789abcdef"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "422": {
                        "description": "Validation error - invalid data or invalid tag IRIs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_tag_iri": {
                                        "summary": "Invalid tag reference",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Invalid tag IRI provided"
                                            }
                                        }
                                    },
                                    "tags_as_objects": {
                                        "summary": "Tags provided as JSON objects instead of IRIs",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Tags must be IRIs, not objects"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/me/bookmarks/{id}": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Get own bookmark by ID",
                "description": "Returns a specific bookmark owned by the authenticated user.",
                "operationId": "getOwnBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Delete a bookmark",
                "description": "Permanently deletes a bookmark (and all its versions/history) owned by the authenticated user.",
                "operationId": "deleteBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bookmark deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Update a bookmark",
                "description": "Updates an existing bookmark. URL cannot be changed after creation (URL field in request body is ignored). Tags are preserved if not included in the request body.",
                "operationId": "updateBookmark",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Bookmark update data",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "New bookmark title",
                                        "type": "string"
                                    },
                                    "url": {
                                        "description": "Ignored - URL cannot be changed after creation",
                                        "type": "string",
                                        "format": "uri",
                                        "deprecated": true
                                    },
                                    "isPublic": {
                                        "description": "Whether the bookmark is public",
                                        "type": "boolean"
                                    },
                                    "tags": {
                                        "description": "Array of tag IRIs. If omitted, existing tags are preserved.",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "iri"
                                        }
                                    },
                                    "mainImage": {
                                        "description": "IRI of main image FileObject",
                                        "type": "string",
                                        "format": "iri",
                                        "nullable": true
                                    },
                                    "archive": {
                                        "description": "IRI of archive FileObject",
                                        "type": "string",
                                        "format": "iri",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPrivate"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required"
                    },
                    "404": {
                        "description": "Bookmark not found or not owned by user"
                    },
                    "422": {
                        "description": "Validation error - invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_data": {
                                        "summary": "Validation error",
                                        "value": {
                                            "error": {
                                                "code": 422,
                                                "message": "Unprocessable Content"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/users/me/files/{id}": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "Get file object (not supported)",
                "description": "This endpoint is not supported. Use POST to upload files.",
                "operationId": "getFileObject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "File object ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Method not allowed - use POST to upload files"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/profile/{username}/bookmarks/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get public bookmark by ID",
                "description": "Returns a specific public bookmark owned by the specified user.",
                "operationId": "getPublicBookmark",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bookmark ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark details. Returns JSON by default (Accept: application/json) or HTML redirect (Accept: text/html).",
                        "headers": {
                            "Location": {
                                "description": "Redirect URL (when Accept: text/html)",
                                "required": false,
                                "schema": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://bookmarkhive.net/profile/username/bookmarks/id"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookmarkShowPublic"
                                }
                            },
                            "text/html": {
                                "schema": {
                                    "description": "HTML redirect response",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "301": {
                        "description": "Redirect response (when Accept: text/html)",
                        "headers": {
                            "Location": {
                                "description": "Redirect URL",
                                "schema": {
                                    "type": "string",
                                    "format": "uri",
                                    "example": "https://bookmarkhive.net/profile/username/bookmarks/id"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bookmark not found, not public, or user not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "description": "Error object",
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "description": "HTTP status code",
                                "type": "integer"
                            },
                            "message": {
                                "description": "Error message",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from POST /auth endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}